pyproject.toml
[project]
name = "discord-amplitude-stats"
version = "0.1.0"
description = "Discord reporting bot for Amplitude analytics and portfolio management"
requires-python = ">=3.11"
dependencies = [
    # Amplitude Report (原有项目)
    "requests>=2.32.0",
    "python-dateutil>=2.9.0",
    
    # Portfolio Report (新项目)
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "numpy-financial>=1.0.0",
    "chinesecalendar>=1.9.0",
]

[project.optional-dependencies]
# Discord Bot 依赖
bot = [
    "discord.py>=2.3.0",
    "openai>=1.0.0",
    "PyGithub>=2.1.0",
]

# 开发依赖
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["portfolio-report", "discord-bot"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # 忽略行长度检查（由 line-length 控制）

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["portfolio-report/core", "portfolio-report/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

